factor out a separate adapter for spans
