+++ exporters/newrelic/src/test/java/io/opentelemetry/exporters/newrelic/NewRelicSpanExporterTest.java
@@ -17,12 +17,11 @@
 package io.opentelemetry.exporters.newrelic;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.verify;
+import static org.mockito.ArgumentMatchers.isA;
 import static org.mockito.Mockito.when;
 
 import com.google.common.primitives.Longs;
 import com.google.protobuf.ByteString;
-import com.google.protobuf.Timestamp;
 import com.newrelic.telemetry.Attributes;
 import com.newrelic.telemetry.exceptions.DiscardBatchException;
 import com.newrelic.telemetry.exceptions.ResponseException;
@@ -31,138 +30,35 @@ import com.newrelic.telemetry.exceptions.RetryWithRequestedWaitException;
 import com.newrelic.telemetry.exceptions.RetryWithSplitException;
 import com.newrelic.telemetry.spans.SpanBatch;
 import com.newrelic.telemetry.spans.SpanBatchSender;
-import io.opentelemetry.proto.trace.v1.AttributeValue;
 import io.opentelemetry.proto.trace.v1.Span;
-import io.opentelemetry.proto.trace.v1.Status;
 import io.opentelemetry.sdk.trace.export.SpanExporter.ResultCode;
-import java.util.Collection;
 import java.util.Collections;
+import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
+import org.mockito.ArgumentMatchers;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
 class NewRelicSpanExporterTest {
 
-  @Mock private SpanBatchSender spanBatchSender;
+  @Mock private SpanBatchSender sender;
+  @Mock private SpanBatchAdapter adapter;
 
   @Test
-  void testSendBatchWithSingleSpan() throws Exception {
-    byte[] spanIdBytes = Longs.toByteArray(1234565L);
-    byte[] bigTraceId = new byte[16];
-    System.arraycopy(Longs.toByteArray(6543215L), 0, bigTraceId, 0, 8);
-    System.arraycopy(Longs.toByteArray(939393939L), 0, bigTraceId, 8, 8);
-    byte[] parentSpanIdBytes = Longs.toByteArray(777666777L);
-
-    com.newrelic.telemetry.spans.Span span1 =
-        com.newrelic.telemetry.spans.Span.builder("000000000012d685")
-            .traceId("000000000063d76f0000000037fe0393")
-            .timestamp(1000456)
-            .name("spanName")
-            .parentId("000000002e5a40d9")
-            .durationMs(1333.020111d)
-            .build();
-    SpanBatch expected = new SpanBatch(Collections.singleton(span1), new Attributes());
-
-    NewRelicSpanExporter testClass = new NewRelicSpanExporter(spanBatchSender, new Attributes());
+  void testExportHappyPath() {
+    NewRelicSpanExporter testClass = new NewRelicSpanExporter(adapter, sender);
 
     Span inputSpan =
-        Span.newBuilder()
-            .setSpanId(ByteString.copyFrom(spanIdBytes))
-            .setTraceId(ByteString.copyFrom(bigTraceId))
-            .setStartTime(Timestamp.newBuilder().setSeconds(1000).setNanos(456_000_000).build())
-            .setParentSpanId(ByteString.copyFrom(parentSpanIdBytes))
-            .setEndTime(Timestamp.newBuilder().setSeconds(1001).setNanos(789_020_111).build())
-            .setName("spanName")
-            .setStatus(Status.newBuilder().setCode(200).build())
-            .build();
-    ResultCode result = testClass.export(Collections.singletonList(inputSpan));
+        Span.newBuilder().setSpanId(ByteString.copyFrom(Longs.toByteArray(1234565L))).build();
+    List<Span> spans = Collections.singletonList(inputSpan);
+    SpanBatch batch =
+        new SpanBatch(Collections.<com.newrelic.telemetry.spans.Span>emptyList(), new Attributes());
+    when(adapter.adaptToSpanBatch(spans)).thenReturn(batch);
 
+    ResultCode result = testClass.export(spans);
     assertEquals(ResultCode.SUCCESS, result);
-    verify(spanBatchSender).sendBatch(expected);
-  }
-
-  @Test
-  void testAttributes() throws Exception {
-    byte[] spanIdBytes = Longs.toByteArray(1234565L);
-    com.newrelic.telemetry.spans.Span span1 =
-        com.newrelic.telemetry.spans.Span.builder("000000000012d685")
-            .timestamp(1000456)
-            .attributes(
-                new Attributes()
-                    .put("myBooleanKey", true)
-                    .put("myIntKey", 123L)
-                    .put("myStringKey", "attrValue")
-                    .put("myDoubleKey", 123.45d))
-            .build();
-    SpanBatch expected = new SpanBatch(Collections.singleton(span1), new Attributes());
-
-    NewRelicSpanExporter testClass = new NewRelicSpanExporter(spanBatchSender, new Attributes());
-
-    Span inputSpan =
-        Span.newBuilder()
-            .setSpanId(ByteString.copyFrom(spanIdBytes))
-            .setStartTime(Timestamp.newBuilder().setSeconds(1000).setNanos(456_000_000).build())
-            .setAttributes(
-                Span.Attributes.newBuilder()
-                    .putAttributeMap(
-                        "myBooleanKey", AttributeValue.newBuilder().setBoolValue(true).build())
-                    .putAttributeMap(
-                        "myIntKey", AttributeValue.newBuilder().setIntValue(123).build())
-                    .putAttributeMap(
-                        "myStringKey",
-                        AttributeValue.newBuilder().setStringValue("attrValue").build())
-                    .putAttributeMap(
-                        "myDoubleKey", AttributeValue.newBuilder().setDoubleValue(123.45d).build()))
-            .build();
-    ResultCode result = testClass.export(Collections.singletonList(inputSpan));
-
-    assertEquals(ResultCode.SUCCESS, result);
-    verify(spanBatchSender).sendBatch(expected);
-  }
-
-  @Test
-  void testMinimalData() throws Exception {
-    byte[] spanIdBytes = Longs.toByteArray(1234565L);
-    NewRelicSpanExporter testClass = new NewRelicSpanExporter(spanBatchSender, new Attributes());
-
-    Span inputSpan = Span.newBuilder().setSpanId(ByteString.copyFrom(spanIdBytes)).build();
-    ResultCode result = testClass.export(Collections.singletonList(inputSpan));
-
-    assertEquals(ResultCode.SUCCESS, result);
-    ArgumentCaptor<SpanBatch> spanBatchCapture = ArgumentCaptor.forClass(SpanBatch.class);
-    verify(spanBatchSender).sendBatch(spanBatchCapture.capture());
-    SpanBatch observedBatch = spanBatchCapture.getAllValues().get(0);
-    Collection<com.newrelic.telemetry.spans.Span> telemetry = observedBatch.getTelemetry();
-    assertEquals(1, telemetry.size());
-  }
-
-  @Test
-  void testErrors() throws Exception {
-    byte[] spanIdBytes = Longs.toByteArray(1234565L);
-    com.newrelic.telemetry.spans.Span span1 =
-        com.newrelic.telemetry.spans.Span.builder("000000000012d685")
-            .timestamp(1000456)
-            .attributes(new Attributes().put("error.message", "it's broken"))
-            .build();
-    SpanBatch expected =
-        new SpanBatch(Collections.singleton(span1), new Attributes().put("host", "localhost"));
-
-    NewRelicSpanExporter testClass =
-        new NewRelicSpanExporter(spanBatchSender, new Attributes().put("host", "localhost"));
-
-    Span inputSpan =
-        Span.newBuilder()
-            .setSpanId(ByteString.copyFrom(spanIdBytes))
-            .setStartTime(Timestamp.newBuilder().setSeconds(1000).setNanos(456_000_000).build())
-            .setStatus(Status.newBuilder().setMessage("it's broken"))
-            .build();
-    ResultCode result = testClass.export(Collections.singletonList(inputSpan));
-
-    assertEquals(ResultCode.SUCCESS, result);
-    verify(spanBatchSender).sendBatch(expected);
   }
 
   @Test
@@ -192,26 +88,18 @@ class NewRelicSpanExporterTest {
   private void checkResponseCodeProducesException(
       ResultCode resultCode, Class<? extends ResponseException> exceptionClass)
       throws ResponseException {
-    byte[] spanIdBytes = Longs.toByteArray(1234565L);
-    com.newrelic.telemetry.spans.Span span1 =
-        com.newrelic.telemetry.spans.Span.builder("000000000012d685")
-            .timestamp(1000456)
-            .attributes(new Attributes().put("error.message", "Some Exception"))
-            .build();
-    SpanBatch expected =
-        new SpanBatch(Collections.singleton(span1), new Attributes().put("host", "localhost"));
+    com.newrelic.telemetry.spans.Span span =
+        com.newrelic.telemetry.spans.Span.builder("000000000012d685").build();
+    SpanBatch spanBatch = new SpanBatch(Collections.singleton(span), new Attributes());
 
-    NewRelicSpanExporter testClass =
-        new NewRelicSpanExporter(spanBatchSender, new Attributes().put("host", "localhost"));
+    NewRelicSpanExporter testClass = new NewRelicSpanExporter(adapter, sender);
 
     Span inputSpan =
-        Span.newBuilder()
-            .setSpanId(ByteString.copyFrom(spanIdBytes))
-            .setStartTime(Timestamp.newBuilder().setSeconds(1000).setNanos(456_000_000).build())
-            .setStatus(Status.newBuilder().setMessage("Some Exception"))
-            .build();
+        Span.newBuilder().setSpanId(ByteString.copyFrom(Longs.toByteArray(1234565L))).build();
 
-    when(spanBatchSender.sendBatch(expected)).thenThrow(exceptionClass);
+    when(adapter.adaptToSpanBatch(ArgumentMatchers.<Span>anyList())).thenReturn(spanBatch);
+    when(sender.sendBatch(isA(SpanBatch.class))).thenThrow(exceptionClass);
+
     ResultCode result = testClass.export(Collections.singletonList(inputSpan));
     assertEquals(resultCode, result);
   }
